+++ src/main/kotlin/icu/windea/pls/cwt/config/CwtDefinitionConfig.kt
@@ -1,24 +1,30 @@
 package icu.windea.pls.cwt.config
 
 import com.intellij.psi.*
+import icu.windea.pls.*
 import icu.windea.pls.cwt.psi.*
-import java.util.WeakHashMap
-
-//TODO 这里排序可能出现问题
+import icu.windea.pls.model.*
+import java.util.*
 
 data class CwtDefinitionConfig(
 	override val pointer: SmartPsiElementPointer<CwtProperty>,
 	val name: String,
-	val config: List<Pair<String?,CwtPropertyConfig>>, //(subtypeExpression, propConfig)
+	val configs: List<Pair<String?,CwtPropertyConfig>> //(subtypeExpression, propConfig)
 ) : CwtConfig<CwtProperty> {
 	//使用WeakHashMap - 减少内存占用
-	private val mergeConfigCache = WeakHashMap<String,List<CwtPropertyConfig>>()
-
-	fun mergeConfig(subtypes: List<String>): List<CwtPropertyConfig> {
+	private val mergeConfigsCache = WeakHashMap<String,List<CwtPropertyConfig>>()
+	private val propertyConfigsCache = WeakHashMap<String,List<CwtPropertyConfig>>()
+	private val childPropertyConfigsCache = WeakHashMap<String,List<CwtPropertyConfig>>()
+	private val childValueConfigsCache = WeakHashMap<String,List<CwtValueConfig>>()
+	
+	/**
+	 * 根据子类型列表合并配置。
+	 */
+	fun mergeConfigs(subtypes: List<String>): List<CwtPropertyConfig> {
 		val cacheKey = subtypes.joinToString(",")
-		return mergeConfigCache.getOrPut(cacheKey){
+		return mergeConfigsCache.getOrPut(cacheKey){
 			val result = mutableListOf<CwtPropertyConfig>()
-			for((subtypeExpression, propConfig) in config) {
+			for((subtypeExpression, propConfig) in configs) {
 				if(subtypeExpression == null || matchesSubtype(subtypeExpression,subtypes)) {
 					result.add(propConfig)
 				}
@@ -27,8 +33,54 @@ data class CwtDefinitionConfig(
 		}
 	}
 	
-	private fun matchesSubtype(subtypeExpression: String, subtypes: List<String>): Boolean {
-		return if(subtypeExpression.startsWith('!')) subtypeExpression.drop(1) !in subtypes else subtypeExpression in subtypes
+	/**
+	 * 根据路径解析对应的属性配置列表。
+	 */
+	fun resolvePropertyConfigs(subtypes: List<String>,path: ParadoxPropertyPath,configGroup: CwtConfigGroup): List<CwtPropertyConfig> {
+		val cacheKey = "${subtypes.joinToString(",")}$path"
+		return propertyConfigsCache.getOrPut(cacheKey){
+			when {
+				//这里的属性路径不应该为空
+				path.isEmpty() -> emptyList()
+				else -> {
+					var result = mergeConfigs(subtypes)
+					var isTop = true
+					for((key, quoted) in path.subPathInfos) {
+						if(isTop) isTop = false else result = result.flatMap { it.properties?: emptyList() }
+						result = result.filter { matchesKey(it.keyExpression, key, quoted, configGroup) }
+					}
+					result
+				}
+			}
+		}
+	}
+	
+	/**
+	 * 根据路径解析对应的子属性配置列表。（过滤重复的）
+	 */
+	fun resolveChildPropertyConfigs(subtypes: List<String>,path: ParadoxPropertyPath, configGroup: CwtConfigGroup): List<CwtPropertyConfig> {
+		val cacheKey = "${subtypes.joinToString(",")}$path"
+		return childPropertyConfigsCache.getOrPut(cacheKey){
+			when {
+				//这里的属性路径可以为空，这时得到的就是顶级属性列表
+				path.isEmpty() -> mergeConfigs(subtypes)
+				else -> resolvePropertyConfigs(subtypes,path, configGroup).flatMap { it.properties?:emptyList() }
+			}
+		}.distinctBy { it.key }
+	}
+	
+	/**
+	 * 根据路径解析对应的子值配置列表。（过滤重复的）
+	 */
+	fun resolveChildValuesConfigs(subtypes: List<String>,path: ParadoxPropertyPath, configGroup: CwtConfigGroup): List<CwtValueConfig> {
+		val cacheKey = "${subtypes.joinToString(",")}$path"
+		return childValueConfigsCache.getOrPut(cacheKey){
+			when {
+				//这里的属性路径可以为空，这时得到的是空列表（假定在顶级的是属性不是值）
+				path.isEmpty() -> emptyList()
+				else -> resolvePropertyConfigs(subtypes,path, configGroup).flatMap { it.values?:emptyList() }
+			}
+		}.distinctBy{it.value}
 	}
 }