+++ src/main/kotlin/icu/windea/pls/model/ParadoxDefinitionInfo.kt
@@ -9,10 +9,11 @@ data class ParadoxDefinitionInfo(
 	val type: String,
 	val typeConfig: CwtTypeConfig,
 	val subtypes: List<String>,
-	val subtypesConfig: List<CwtSubtypeConfig>,
+	val subtypeConfigs: List<CwtSubtypeConfig>,
 	val localisation: List<ParadoxDefinitionLocalisationInfo>,
-	val localisationConfig: List<CwtTypeLocalisationConfig>,
-	val definitionConfig: List<CwtPropertyConfig>,
+	val localisationConfig: CwtTypeLocalisationConfig?,
+	val definition: List<CwtPropertyConfig>,
+	val definitionConfig:CwtDefinitionConfig?,
 	val typeKey: String,
 	val graphRelatedTypes: List<String>,
 	val unique: Boolean,
@@ -31,15 +32,11 @@ data class ParadoxDefinitionInfo(
 			appendPsiLink("$typeLink.$subtype", subtype)
 		}
 	}
-	val typePointer = typeConfig.pointer
-	val subtypesPointer = subtypesConfig.map { it.pointer }
-	val typesPointer = mutableListOf(typePointer).apply { addAll(subtypesPointer) }
+	//val typePointer = typeConfig.pointer
+	//val subtypesPointer = subtypesConfig.map { it.pointer }
+	//val typesPointer = mutableListOf(typePointer).apply { addAll(subtypesPointer) }
 	val localisationNames = localisation.map { it.name }
 	val localisationKeyNames = localisation.map { it.keyName }
-	
-	private val propertyConfigsCache = WeakHashMap<String,List<CwtPropertyConfig>>()
-	private val childPropertyConfigsCache = WeakHashMap<String,List<CwtPropertyConfig>>()
-	private val childValueConfigsCache = WeakHashMap<String,List<CwtValueConfig>>()
 	
 	override fun equals(other: Any?): Boolean {
 		return this === other || other is ParadoxDefinitionInfo && name == other.name && types == other.types
@@ -63,49 +60,20 @@ data class ParadoxDefinitionInfo(
 	 * 根据路径解析对应的属性配置列表。
 	 */
 	fun resolvePropertyConfigs(path: ParadoxPropertyPath,configGroup: CwtConfigGroup): List<CwtPropertyConfig> {
-		val cacheKey = path.toString()
-		return propertyConfigsCache.getOrPut(cacheKey){
-			when {
-				//这里的属性路径不应该为空
-				path.isEmpty() -> emptyList() 
-				else -> {
-					var result = definitionConfig
-					var isTop = true
-					for((key, quoted) in path.subPathInfos) {
-						if(isTop) isTop = false else result = result.flatMap { it.properties?: emptyList() }
-						result = result.filter { matchesKey(it.keyExpression, key, quoted, configGroup) }
-					}
-					result
-				}
-			}
-		}
+		return definitionConfig?.resolvePropertyConfigs(subtypes,path,configGroup)?: emptyList()
 	}
 	
 	/**
 	 * 根据路径解析对应的子属性配置列表。（过滤重复的）
 	 */
 	fun resolveChildPropertyConfigs(path: ParadoxPropertyPath, configGroup: CwtConfigGroup): List<CwtPropertyConfig> {
-		val cacheKey = path.toString()
-		return childPropertyConfigsCache.getOrPut(cacheKey){
-			when {
-				//这里的属性路径可以为空，这时得到的就是顶级属性列表
-				path.isEmpty() -> definitionConfig
-				else -> resolvePropertyConfigs(path, configGroup).flatMap { it.properties?:emptyList() }
-			}
-		}.distinctBy { it.key }
+		return definitionConfig?.resolveChildPropertyConfigs(subtypes,path,configGroup)?: emptyList()
 	}
 	
 	/**
 	 * 根据路径解析对应的子值配置列表。（过滤重复的）
 	 */
 	fun resolveChildValuesConfigs(path: ParadoxPropertyPath, configGroup: CwtConfigGroup): List<CwtValueConfig> {
-		val cacheKey = path.toString()
-		return childValueConfigsCache.getOrPut(cacheKey){
-			when {
-				//这里的属性路径可以为空，这时得到的是空列表（假定在顶级的是属性不是值）
-				path.isEmpty() -> emptyList() 
-				else -> resolvePropertyConfigs(path, configGroup).flatMap { it.values?:emptyList() }
-			}
-		}.distinctBy{it.value}
+		return definitionConfig?.resolveChildValuesConfigs(subtypes, path, configGroup)?: emptyList()
 	}
 }