+++ src/main/kotlin/icu/windea/pls/model/ParadoxLocalisationInfo.kt
@@ -13,8 +13,4 @@ data class ParadoxLocalisationInfo(
 	override fun hashCode(): Int {
 		return Objects.hash(name, category)
 	}
-	
-	override fun toString(): String {
-		return "(${category.key}) $name"
-	}
 }